Scaffolding
This commit is a bunch of scaffolding building on the idea of a distributed workflow
manager similar to cadence or temporal. It contains protobuf definitions for a controller
and worker that communicate both ways in order to run workflows and their individual
tasks.

Included is a basic scheduler implementation and the bare bones of the API used to
write a workflow and tasks.

Signed-off-by: David Bond <davidsbond93@gmail.com>