It lives!
This commit adds all the gubbins required to get the gRPC servers up and running
and to perform registration/deregistration of the workers with the controllers.

Up next, we'll need the async process for the controllers to pick up pending
tasks/workflows and yeet them to the workers that support them.

Signed-off-by: David Bond <davidsbond93@gmail.com>