We have liftoff
This commit gets everything into a state where I was able to run a workflow with a task
from the CLI. The controller and worker now both communicate and successfully run
the task.

Included is a separation between the public and private APIs for the CLI and controller/worker
communication respectively. A few fixes to some SQL queries and generally how we store
inputs and outputs.

Signed-off-by: David Bond <davidsbond93@gmail.com>