Let there be child workflows
This commit modifies the workflow API to allow child workflows to be spawned from a
workflow. To make this work the worker had to be able to manage more than one
workflow/task happening at a time, which it does now! This is hard-coded to 10
each for the time being but should probably be tuneable at some point.

Signed-off-by: David Bond <davidsbond93@gmail.com>