The tree must grow
This commit adds the `orca workflow describe` command that will call the new describe endpoint
and provides two options for output. The one I like is the default tree based output that looks
like this:

```
Workflow: TestWorkflow (3.457s) [COMPLETE]
└── Task: TestTask (1.437s) [COMPLETE]
└── Workflow: TestChildWorkflow (2.002s) [COMPLETE]
    └── Task: TestTask (1.989s) [COMPLETE]
```

Otherwise, a `--json` flag can be provided to just output the whole description as a JSON
object.

Signed-off-by: David Bond <davidsbond93@gmail.com>