Add task and workflow views
This commit adds materialized views for workflow and task names so they can be
queried to see what we've got. This is powered by a trigger that refreshes
the materialized view whenever a worker joins or leaves.

This can be used in future to prevent scheduling with a bad workflow or task
name.

Signed-off-by: David Bond <davidsbond93@gmail.com>