Idempotent keys
This commit adds the concept of idempotent keys to child workflows and tasks. These keys can
be used to skip workflows/tasks that have already been run based on their inputs. This is
an initial approach to handling workflow failing and retries.

If we want to retry a workflow, we need to make sure we don't rerun tasks whose actions
are not explicitly idempotent. To avoid all task code having to handle things already
existing the user can provide an idempotent key.

If a task/workflow is scheduled with one of these keys, the database is checked for an
existing run with a COMPLETED status that matches the workfow/task name and key. If it
exists, this new run is marked as skipped and won't be scheduled.

The worker will then treat statuses of COMPLETED and SKIPPED the same.

Signed-off-by: David Bond <davidsbond93@gmail.com>